requirements.txt
python-dotenv==1.0.0
aiogram==2.25.1
sqlalchemy==2.0.19
asyncpg==0.28.0
google-api-python-client==2.94.0
google-api-core==2.11.1
google-auth==2.22.0
googleapis-common-protos==1.59.1
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
